Change how debug switch is implemented
As we add new test cases, we needed something more
targeted than a global switch. This introduce a local
method (package private) that let us control on a per
mock basis when to enable debug logs. If need be, we
could make it public in the `mock` function. For now
it's considered an implementation detail and thus
hidden